% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_function.R
\name{w}
\alias{w}
\title{How many hours, years, or centuries have you spent
meticulously writing out
"write_csv(whatever, 'whatever.csv')"
perhaps followed by "view(whatever)"?
No more.}
\usage{
w(df = NULL, view = F)
}
\description{
Arguments:
\enumerate{
\item df = the dataframe you want to write as a csv to your working directory. Defaults to NULL
\item v (for "view") defaults to FALSE, but make it T to have the df automatically show up in the viewer after writing
\item There is no 3, because I don't really know what I'm doing
}
}
\details{
Example:

GOATS = tibble(name=c("Juice","Olivia","Mac","Billie"), alive=c(F, T, F, T))

w(GOATS) # new CSV just dropped

w(GOATS,T) #same but now you see it too

GOATS\%>\%w(t) #This way doesn't work but idk why  ¯\(ツ)_/¯

Thanks to the following website for enabling this monstrosity:

https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/
}
